Fixed the DIRECTORY-FN for SBCL.
Note that this may not be the right thing, although it works.  The
problem is that the file-browser context populates itself with a
pathname based on parsing a directory name.  The problem is that for
DIRECTORY to work on a pathname in SBCL, the pathname should have
:name :unspecific and :type similarly.  I kludged this by sticking a
"*" on in the directory-fn (which operates on the namestring), but
perhaps getting the pathname set properly would be better.